requirements.txt
beautifulsoup4==4.6.0
Django==1.11.8
django-extensions==1.9.7
djangorestframework==3.7.3
lxml==4.1.1
requests==2.18.4
